requirements.txt
playwright>=1.43.0
python-dateutil>=2.9.0
